+++ plugin/quotesplugin/plugin_test.go
@@ -1,0 +1,88 @@
+package quotesplugin
+
+import (
+	"testing"
+
+	"git.abyle.org/redseligg/botorchestrator/botconfig"
+	"github.com/stretchr/testify/assert"
+	"github.com/torlenor/abylebotter/model"
+	"github.com/torlenor/abylebotter/plugin"
+)
+
+func TestCreateQuotesPlugin(t *testing.T) {
+	assert := assert.New(t)
+
+	p, err := New(botconfig.PluginConfig{Type: "something"})
+	assert.Error(err)
+	assert.Nil(p)
+
+	p, err = New(botconfig.PluginConfig{Type: PLUGIN_TYPE})
+	assert.NoError(err)
+	assert.Equal(nil, p.API)
+
+	api := plugin.MockAPI{}
+	storage := plugin.MockStorageAPI{}
+	p.SetAPI(&api, &storage)
+}
+
+func TestQuotesPlugin_HelpTextAndInvalidCommands(t *testing.T) {
+	assert := assert.New(t)
+
+	p, err := New(botconfig.PluginConfig{Type: PLUGIN_TYPE})
+	assert.NoError(err)
+	assert.Equal(nil, p.API)
+
+	api := plugin.MockAPI{}
+	storage := plugin.MockStorageAPI{}
+	p.SetAPI(&api, &storage)
+
+	postToPlugin := model.Post{
+		ChannelID: "CHANNEL ID",
+		Channel:   "SOME CHANNEL",
+		User:      model.User{ID: "SOME USER ID", Name: "USER 1"},
+		Content:   "MESSAGE CONTENT",
+		IsPrivate: false,
+	}
+	p.OnPost(postToPlugin)
+	assert.Equal(false, api.WasCreatePostCalled)
+
+	api.Reset()
+	postToPlugin.Content = "!quoteadd"
+	expectedPostFromPlugin := model.Post{
+		ChannelID: "CHANNEL ID",
+		Content:   helpText,
+		IsPrivate: false,
+	}
+	p.OnPost(postToPlugin)
+	assert.Equal(true, api.WasCreatePostCalled)
+	assert.Equal(expectedPostFromPlugin, api.LastCreatePostPost)
+
+	api.Reset()
+	postToPlugin.Content = "!quoteadd "
+	p.OnPost(postToPlugin)
+	assert.Equal(true, api.WasCreatePostCalled)
+	assert.Equal(expectedPostFromPlugin, api.LastCreatePostPost)
+
+	api.Reset()
+	postToPlugin.Content = "!quotehelp"
+	p.OnPost(postToPlugin)
+	assert.Equal(true, api.WasCreatePostCalled)
+	assert.Equal(expectedPostFromPlugin, api.LastCreatePostPost)
+
+	api.Reset()
+	postToPlugin.Content = "!quoteremove"
+	expectedPostFromPlugin = model.Post{
+		ChannelID: "CHANNEL ID",
+		Content:   helpTextRemove,
+		IsPrivate: false,
+	}
+	p.OnPost(postToPlugin)
+	assert.Equal(true, api.WasCreatePostCalled)
+	assert.Equal(expectedPostFromPlugin, api.LastCreatePostPost)
+
+	api.Reset()
+	postToPlugin.Content = "!quoteremove something"
+	p.OnPost(postToPlugin)
+	assert.Equal(true, api.WasCreatePostCalled)
+	assert.Equal(expectedPostFromPlugin, api.LastCreatePostPost)
+}