+++ plugin/quotesplugin/quote.go
@@ -1,0 +1,10 @@
+package quotesplugin
+
+type quote struct {
+	Author   string
+	AuthorID string
+
+	ChannelID string
+
+	Text string
+}